+++ src/Components/App/App.js
@@ -3,41 +3,37 @@ import { Route, Routes } from "react-router-dom";
 import React from "react";
 import Login from "../Login/Login";
 import "./App.css";
+import { fetchToken } from "../../fetchcalls";
+import spotifyLogo from "../../Spotify_Logo_RGB_Green.png";
 
 function App() {
   const clientID = process.env.REACT_APP_API_ID;
-  const clientSecret = process.env.REACT_APP_API_SECRET ;
+  const clientSecret = process.env.REACT_APP_API_SECRET;
   const [token, setToken] = useState(null);
-  
-  useEffect(() => {
-    const requestOptions = {
-      method: "POST",
-      headers: {
-        Authorization: "Basic " + btoa(`${clientID}:${clientSecret}`),
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-      body: "grant_type=client_credentials",
-    };
 
-    fetch("https://accounts.spotify.com/api/token", requestOptions)
-      .then((response) => response.json())
-      .then((data) => {
-        if (data.access_token) {
-          setToken(data.access_token);
-        }
+  useEffect(() => {
+    fetchToken(clientID, clientSecret)
+      .then((accessToken) => {
+        setToken("Access Token:", accessToken);
       })
-      .catch((error) => console.error("Error:", error));
+      .catch((error) => {
+        console.error("Error:", error);
+      });
   }, []);
 
   return (
-    <div>
-      <p>Access Token: {token}</p>
-      <p>Client Id: {clientID}</p>
-      <Login clientId={clientID} />
-      <Routes>
-        <Route path="/dashboard" element={<p>YOU DID IT</p>} />
-      </Routes>
-    </div>
+    <>
+      <header>
+        <h1>Playlist Builder</h1>
+        <img className="spotifyLogo" src={spotifyLogo} />
+      </header>
+      <main>
+        <Routes>
+          <Route path="/" element={<Login clientId={clientID} />} />
+          <Route path="/dashboard" element={<p>YOU DID IT</p>} />
+        </Routes>
+      </main>
+    </>
   );
 }
