+++ src/Components/Login/Login.css
@@ -1,0 +1,12 @@
+.loginBtn {
+    position: absolute;
+    left: 40%;
+    top:45%;
+    width: 300px;
+    height: 75px;
+    background-color: black;
+    color: white;
+    border:none;
+    border-radius: 2%;
+    font-size: larger;
+}
